+++ strkey/main.go
@@ -189,3 +189,29 @@ func decodeString(src string) ([]byte, error) {
 
 	return raw, nil
 }
+
+// IsValidEd25519PublicKey validates a stellar public key
+func IsValidEd25519PublicKey(i interface{}) bool {
+	enc, ok := i.(string)
+
+	if !ok {
+		return false
+	}
+
+	_, err := Decode(VersionByteAccountID, enc)
+
+	return err == nil
+}
+
+// IsValidEd25519SecretSeed validates a stellar secret key
+func IsValidEd25519SecretSeed(i interface{}) bool {
+	enc, ok := i.(string)
+
+	if !ok {
+		return false
+	}
+
+	_, err := Decode(VersionByteSeed, enc)
+
+	return err == nil
+}