+++ txnbuild/inflation.go
@@ -22,3 +22,12 @@ func (inf *Inflation) BuildXDR() (xdr.Operation, error) {
 	SetOpSourceAccount(&op, inf.SourceAccount)
 	return op, nil
 }
+
+// FromXDR for Inflation initialises the txnbuild struct from the corresponding xdr Operation.
+func (inf *Inflation) FromXDR(xdrOp xdr.Operation) error {
+	if xdrOp.Body.Type != xdr.OperationTypeInflation {
+		return errors.New("error parsing inflation operation from xdr")
+	}
+	inf.SourceAccount = accountFromXDR(xdrOp.SourceAccount)
+	return nil
+}